requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
Pillow>=8.3.0
requests>=2.26.0

# Optional: for GPU acceleration (if available)
# torch>=1.9.0

# Development dependencies (optional)
# pytest>=6.2.0  # Alternative test runner
# jupyter>=1.0.0  # For interactive exploration